requirements.txt
astroid==2.9.0
autopep8==1.6.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.12.2
pyserial==3.5
toml==0.10.2
typing_extensions==4.0.1
wrapt==1.13.3
yamspy==0.3.3
